🚚 indexページを作成して既存のものをhome.jsに移動
