+++ src/pages/home.js
@@ -1,0 +1,23 @@
+import Head from 'next/head'
+import { Inter } from 'next/font/google'
+import Header from '@/components/Header'
+import Menu from '@/components/Menu'
+import HomeContent from '@/components/HomeContent'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export default function Home() {
+  return (
+    <>
+    <Head>
+      <title>wakaru | Home</title>
+      <meta name="description" content="Generated by create next app" />
+      <meta name="viewport" content="width=device-width, initial-scale=1" />
+      <link rel="icon" href="/favicon.ico" />
+    </Head>
+    <Header />
+    <HomeContent />
+    <Menu />
+    </>
+  )
+}
