+++ src/pages/index.js
@@ -3,21 +3,17 @@ import { Inter } from 'next/font/google'
 import Header from '@/components/Header'
 import Menu from '@/components/Menu'
 import HomeContent from '@/components/HomeContent'
+import Link from 'next/link'
+import styles from '@/styles/Home.module.css'
 
 const inter = Inter({ subsets: ['latin'] })
 
-export default function Home() {
+export default function Index() {
   return (
-    <>
-    <Head>
-      <title>wakaru | Home</title>
-      <meta name="description" content="Generated by create next app" />
-      <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <link rel="icon" href="/favicon.ico" />
-    </Head>
-    <Header />
-    <HomeContent />
-    <Menu />
-    </>
+    <div className={styles.main}>
+      <div>こんなアプリですよ～</div>
+      <hr />
+      <Link href="home">使ってみる→</Link>
+    </div>
   )
 }
